[incfs] Allow taking out the memory mapping from map_ptr
The Android resources code needs to be able to take over the
internal memory mapping to efficiently parse the data inside.

This CL replaces the outdated FileMap class with MappedFile,
which allows moving it out efficiently. This also gets rid of the
libutils dependency that was only there for FileMap

Bug: b/443273051
Test: atest libincfs-test
Flag: android.content.res.xml_file_size_limit
Change-Id: I5e77aa7137357fde7c86612300b0033a800ddee7